[feat] add genral folder in components
