--- MyMovie/src/components/PrimaryBtn.vue
@@ -1,22 +1,0 @@
-<!-- why padding doesn't work !!! -->
-<template>
-    <a
-    href="#"
-    class="bg-primary rounded-xl inline-block text-white shadow-lg shadow-primary-dark text-sm"
-    :style="{paddingTop : py, paddingBottom: py, paddingRight: px, paddingLeft: px}"
-    ><slot>More Detail</slot></a
-    >
-</template>
-
-<script setup>
-    defineProps({
-        'px' : {
-            type: String,
-            default : '16px'
-        },
-        'py' : {
-            type: String,
-            default : '12px'
-        }
-    })
-</script>
