+++ MyMovie/src/components/SuggestionBox.vue
@@ -26,8 +26,8 @@
 </template>
 
 <script setup>
-import AddToWatchListBtn from './AddToWatchListBtn.vue';
-import PrimaryBtn from './PrimaryBtn.vue';
+import AddToWatchListBtn from './general/AddToWatchListBtn.vue';
+import PrimaryBtn from './general/PrimaryBtn.vue';
 defineProps([
   'backgroundImage',
   'title',
